many changes. bkp
