+++ .github/workflows/main.yml
@@ -35,7 +35,7 @@ jobs:
     - run: npm run build
 
     # Create a zip file with all files required by the module to add to the release
-    - run: zip -r ./module.zip module.json readme.md LICENSE css/ templates/ utils/ *.mjs sample-root-soundscapes/ logo.webp
+    - run: zip -r ./module.zip module.json readme.md LICENSE css/ templates/ utils/ scripts/ sample-root-soundscapes/ logo.webp
 
     # Create a release for this specific version
     - name: Update Release with Files
