+++ lang/en.json
@@ -1,0 +1,26 @@
+{
+    "TYPES.RegionBehavior": {
+      "soundscape-adventure.moodPlay": "Mood Play",
+      "soundscape-adventure.moodPlayEvents": "Mood Play Events"
+    },
+    "RAE.ACTIONS": {
+      "add.label": "Add",
+      "resetDuration.label": "Reset Duration",
+      "enable.label": "Enable",
+      "disable.label": "Disable",
+      "delete.label": "Delete"
+    },
+    "RAE.TYPES.moodPlay": {
+      "FIELDS.moodId.label": "Mood Name",
+      "FIELDS.moodId.hint": "The Mood Play to apply when one of the events occurs.",
+      "FIELDS.overlay.label": "Overlay",
+      "FIELDS.overlay.hint": "Display the Mood Play as an overlay."
+    },
+    "RAE.TYPES.moodPlayEvents": {
+      "FIELDS.action.choices.custom": "Play (custom)",
+      "FIELDS.action.choices.play": "Play",
+      "FIELDS.action.choices.stop": "Stop",
+      "FIELDS.action.label": "Action",
+      "FIELDS.action.hint": "What to do with the Mood Play."
+    }
+  }
