+++ scripts/utils/constants.mjs
@@ -1,0 +1,53 @@
+const LOGLEVEL = {
+    INFO: 0,
+    WARN: 1,
+    ERROR: 2
+}
+
+const MOOD = {
+    STATUS: {
+        STOP: 1,
+        PLAYING: 2
+    }
+}
+
+const SOUNDTYPE = {
+    AMBIENCE: 0,
+    LOOP: 1,
+    RANDOM: 2,
+    SOUNDPAD: 3,
+    GROUP_LOOP: 4,
+    GROUP_RANDOM: 5,
+    GROUP_SOUNDPAD: 6,
+    INVALID: -1
+
+}
+
+const SOUNDSCAPE_TYPE = {
+    LOCAL: 0,
+    REMOTE_S3: 1
+}
+
+const PREFIX="Soundscape";
+
+const STORAGETRIGGERSETTINGS = "soundscape-adventure.triggerSettings";
+/* trigger config example
+const triggerData = {
+  type: SOUND | GROUP | THEME,
+  triggerScope: "scene",
+  triggerName: "",
+  
+};
+
+*/
+
+
+export default {
+    LOGLEVEL,
+    MOOD,
+    SOUNDTYPE,
+    SOUNDSCAPE_TYPE,
+    PREFIX,
+    STORAGETRIGGERSETTINGS
+    // Add other constants here
+  };
