+++ templates/editmood.hbs
@@ -1,0 +1,54 @@
+<div class="container">
+    <h1>Mood Settings: {{ mood.name }}</h1>
+
+        <!-- Mood Name -->
+        <div class="form-group">
+            <label for="moodName">Mood Name:</label>
+            <input type="text" id="moodName" name="moodName" placeholder="Enter mood name" value="{{ mood.name }}"/>
+        </div>
+        <br />
+        <!-- Trigger -->
+        <h3>Triggers<i class="fa-solid fa-block-question" data-tooltip="For region triggers you need to go to the region's behaviour"></i></h3>
+        <div class="form-group inline-group">
+            {{#each triggers}}
+            <div class="trigger_group">
+                <h4>Trigger {{ @index }}</h4>
+                <div class="form-group">
+                    <label for="triggerAction{{ @index }}">Action</label>
+                    <select id="triggerAction{{ @index }}" name="triggerAction">
+                        <option value="">Select</option>
+                        {{#each this.actions}}
+                            <option value="{{ this.id }}" {{#if this.selected }}selected {{/if}}>{{ this.name }}</option>
+                        {{/each}}
+                    </select>
+                </div>
+                <br />
+                <div class="form-group">
+                    <label for="triggerOn{{ @index }}">On</label>
+                    <select class="onElement" id="triggerOn{{ @index }}" name="triggerOn">
+                        <option value="">Select</option>
+                        {{#each this.on}}
+                            <option value="{{ this.id }}" {{#if this.selected }}selected {{/if}}>{{ this.name }}</option>
+                        {{/each}}
+                    </select>
+                </div>
+                <br />
+                <div class="form-group">
+                    <label for="triggerEvent{{ @index }}">Event</label>
+                    <select class="eventElement" id="triggerEvent{{ @index }}" name="triggerEvent">
+                        <option value="">Select</option>
+                        {{#each this.events}}
+                            <option value="{{ this.id }}" {{#if this.selected }}selected {{/if}}>{{ this.name }}</option>
+                        {{/each}}
+                    </select>
+                </div>
+            </div> <!-- trigger group -->
+            {{/each}}
+        </div>
+        <!-- # Region Triggers -->
+        <br />
+        <br />
+        <button class="removeAllTriggers" data-mood-id="{{ mood.id }}">
+            <i class="fa-solid fa-broom"></i>Remove all triggers for this mood
+        </button>
+</div>
